requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord==1.0.1
discord.py==1.5.0
idna==2.9
multidict==4.7.6
mysql-connector-python==8.0.21
numpy==1.18.5
Pillow==7.1.2
protobuf==3.12.2
python-decouple==3.3
pytz==2020.1
six==1.15.0
websockets==8.1
yarl==1.4.2